添加uD增磁/弱磁分析测试和Streamlit默认参数调整
新增测试:
- tests/test_ud_analysis.py: uD从-5V到+5V扫描, 验证弱磁升速和增磁降速
- tests/test_ud_uq_analysis.py: uD/uQ组合分析(Playwright版本)

St_main参数调整:
- 默认控制周期100us改为50us
- 默认仿真步数100000改为4000
- uD打印格式修正(%ld替代%.1f)

+++ frameworkCodes/c/main_velocity_test.c
@@ -249,7 +249,7 @@ int main(void) {
     printf("=== Electric Machinery Simulation ===\n");
     printf("Mode: %ld\n", d_sim.simulation.mode_select);
     printf("  cmd_speed = %.1f rad/s\n", d_sim.test.cmd_speed);
-    printf("  cmd_uD = %.1f V, cmd_uQ = %.1f V\n", d_sim.test.cmd_ud, d_sim.test.cmd_uq);
+    printf("  cmd_uD = %ld V, cmd_uQ = %.1f V\n", (long)d_sim.test.cmd_ud, d_sim.test.cmd_uq);
     printf("  TLoad = %.4f Nm\n", d_sim.test.tload);
     printf("\n");
